TODO: Fix collision with tiles! TODO: Change collision to check only when entities attempt to move! This will allow us to disable movement when encountering a solid entity.
Changes:
- Castle:
* Now spawns health crystals.
* Changed the tower levels to be non-collision entities.
 *Castle base now renders behind all other entities. Tower levels render above.
 - Mob:
 * Added maxHP, and also public method to modify Mob's health.
 - Player:
 * Set maxHP to 99.
 - PowerUp:
 * Created PowerUp abstract class for pickup objects.
 - HealthCrystal:
 * Added healing crystal with ballistics!
 - CastleFloor:
 * Added CastleFloor entity to represent the tower's various levels.
 - Entity:
 * Modified X,Y coordinates to be double, to allow for fractional speeds.
 - Sprite + Screen + SpriteSheet + TiledSpriteSheet + AnimatedSprite:
 * Allow for non-square sprites!

** NOTE: MOB coordinates (x,y) are of the CENTER of the sprite! This is why there are many awkward adjustments of sprite.HEIGHT/2 ETC. May want to change this!

+++ src/com/arena/entity/CollidableEntity.java
@@ -39,6 +39,6 @@ public abstract class CollidableEntity extends Entity {
     }
 
     public BoundingBox getCollisionBox(){
-        return sprite.getCollisionBox().translate(x - sprite.WIDTH/2,y-sprite.HEIGHT/2);
+        return sprite.getCollisionBox().translate((int)x - sprite.WIDTH/2,(int)y-sprite.HEIGHT/2);
     }
 }
