+++ src/com/arena/entity/Entity.java
@@ -2,12 +2,13 @@ package com.arena.entity;
 
 import com.arena.graphics.Screen;
 import com.arena.GameScreen.level.Level;
+import com.arena.graphics.Sprite;
 
 import java.util.Random;
 
 public abstract class Entity {
 
-    public int x, y;
+    public double x, y;
     protected boolean removed = false;
     protected Level level;
     protected final Random random = new Random();
@@ -18,12 +19,14 @@ public abstract class Entity {
         UP;
     }
 
+
     public void update(){
     }
 
     public void render(Screen screen){
 
     }
+
 
     public void init(Level level){
         this.level = level;
